Update and split interfaces
