+++ src/WmiScripting/Contracts/CastsAttributes.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PhpWinTools\WmiScripting\Contracts;
+
+interface CastsAttributes extends HasAttributes
+{
+    public function getCasts(): array;
+
+    public function getCast($attribute);
+}
