+++ src/WmiScripting/Contracts/HasAttributes.php
@@ -5,14 +5,4 @@ namespace PhpWinTools\WmiScripting\Contracts;
 interface HasAttributes extends Arrayable
 {
     public function getAttribute($attribute, $default = null);
-
-    public function mergeHiddenAttributes(array $hidden_attributes, bool $merge_hidden = true);
-
-    public function getHiddenAttributes();
-
-    public function isHidden($attribute): bool;
-
-    public function getCasts(): array;
-
-    public function getCast($attribute);
 }
