+++ src/WmiScripting/Models/Win32Model.php
@@ -8,6 +8,8 @@ use PhpWinTools\WmiScripting\Contracts\Jsonable;
 use PhpWinTools\WmiScripting\Contracts\Arrayable;
 use PhpWinTools\WmiScripting\Contracts\HasAttributes;
 use PhpWinTools\WmiScripting\MappingStrings\Mappings;
+use PhpWinTools\WmiScripting\Contracts\CastsAttributes;
+use PhpWinTools\WmiScripting\Contracts\HidesAttributes;
 use function PhpWinTools\WmiScripting\Support\connection;
 use PhpWinTools\WmiScripting\Collections\ModelCollection;
 use PhpWinTools\WmiScripting\Concerns\HasHiddenAttributes;
@@ -22,7 +24,7 @@ use PhpWinTools\WmiScripting\Support\ApiObjects\Contracts\ObjectPath;
 /**
  * @link https://docs.microsoft.com/en-us/windows/win32/cimwin32prov/win32-provider
  */
-class Win32Model implements Arrayable, Jsonable, HasAttributes
+class Win32Model implements Arrayable, Jsonable, HasAttributes, HidesAttributes, CastsAttributes
 {
     use HasArrayableAttributes,
         HasHiddenAttributes,
